network stuff moved to own module
